update headings & fix book error (see desc)
In the book, section ## IV. Propagation of the... comes before ###3. The hearing lineage of...

HOWERVER

That doesnt seem correct based on 2 observations. The 1st being the information architecture hierarchy itself breaks as there are no ###1 and ###2 before the ###3 now.

The second way it breaks is contextually it doesnt make sense to remove the 3rd lineage from its list